cloud_storage: Remove assertion from recovery code
The assertion in the recovery process is triggered when the manifest
upload fails. Previousy, this was critical because the manifest was
downloaded after first leader election. This manifest was then used to
upload new data. If the manifest wasn't uploaded the newly created
ntp_archiver will 'see' that the manifest is missing. This means that
the partition wasn't archived before. After that it should start
uploading all the data in the log. This will lead to re-upload of all
data in the partition.

Currently, this is not a problem. Recovery creates an
archival_metadata_stm snapshot. This snapshot contains same data as the
manifest. The manifest is re-uploaded periodically, so it's safe to just
log this error.